ML/requirements.txt
fastapi
uvicorn
python-multipart
torch --index-url https://download.pytorch.org/whl/cpu
open-clip-torch
faiss-cpu
numpy
Pillow